# render: add alpha support for filtering