--- a/alpha/declcfg/filter/config/v1alpha1/testdata/configs/invalid_apiversion.yaml
+++ b/alpha/declcfg/filter/config/v1alpha1/testdata/configs/invalid_apiversion.yaml
@@ -1,0 +1,10 @@
+apiVersion: olmfiltering.operatorframework.io/v1alpha1
+kind: FilterConfiguration
+packages:
+  - name: "foo"
+  - name: "bar"
+    channels:
+      - name: "bar-channel1"
+        versionRange: ">=1.0.0 <2.0.0"
+      - name: "bar-channel2"
+        versionRange: ">=2.0.0 <3.0.0"