--- a/alpha/declcfg/filter/config/v1alpha1/testdata/configs/invalid_missingpackagename.yaml
+++ b/alpha/declcfg/filter/config/v1alpha1/testdata/configs/invalid_missingpackagename.yaml
@@ -1,0 +1,5 @@
+apiVersion: olm.operatorframework.io/v1alpha1
+kind: FilterConfiguration
+packages:
+  - name: "foo"
+  - name: ""