--- a/go.mod
+++ b/go.mod
@@ -3,6 +3,7 @@ module github.com/operator-framework/operator-registry
 go 1.21
 
 require (
+	github.com/Masterminds/semver/v3 v3.2.1
 	github.com/blang/semver/v4 v4.0.0
 	github.com/containerd/containerd v1.7.14
 	github.com/containers/common v0.58.1