… alpha/action/render_test.go
… alpha/declcfg/filter/config/v1alpha1/channel_test.go
… alpha/declcfg/filter/config/v1alpha1/config_test.go
… alpha/declcfg/filter/config/v1alpha1/filter_test.go
… alpha/declcfg/filter/config/v1alpha1/testdata/configs/invalid_kind.yaml
… alpha/declcfg/filter/config/v1alpha1/testdata/configs/invalid_missingchannelname.yaml
… alpha/declcfg/filter/config/v1alpha1/testdata/configs/invalid_multipleerrors.yaml
… alpha/declcfg/filter/config/v1alpha1/testdata/configs/invalid_nopackages.yaml
… alpha/declcfg/filter/config/v1alpha1/testdata/configs/valid.yaml
… alpha/declcfg/filter/filter.go
… alpha/declcfg/filter/filter_test.go
… alpha/declcfg/load_test.go
… alpha/property/scheme_test.go
… cmd/opm/render/cmd.go
… go.sum