<!--

Before making a PR, please read our contributing guidelines https://github.com/operator-framework/operator-lifecycle-manager/blob/master/CONTRIBUTING.md
Note: Make sure your branch is rebased to the latest upstream master.

-->

**Description of the change:**

This commit adds two alpha flags to `opm render` and updates FBC library code to support efficient filtering of FBC:
- `--alpha-filter-config=./path/to/config`
- `--alpha-keep-packages=pkg1,pkg2`

To support the filter config option, this commits defines a new configuration file format that supports:
- filtering by package
- within a package, filtering by channel
- within a package, selecting a new default channel (required if the original default channel is filtered out)
- within each channel, filtering by a version range

In order to support correct version range filtering, some bundles that fall outside the version range may remain in the channel to ensure that the filtered channel is still valid and includes all bundles that did fall in the range

Included in this commit is an update to declcfg.WalkMetasFS to enable concurrency, similar to recent changes in declcfg.LoadFS. declcfg.LoadFS is also refactored internally to use the new declcfg.WalkMetasFS implementation.

**Motivation for the change:**

There are two motivations:
1. Filtering by individual versions is an often requested feature, especially in workflows that requiring mirroring catalogs to disconnected environments. In OLM, filtering versions requires semantic knowledge of the upgrade graph rules, so it is not as simple as keeping or removing arbitrary bundles and channel entries. This PR includes that semantic knowledge.
2. When rendering an existing FBC, `render` is significantly faster than rendering an entire catalog and then using a post-render tool like jq or yq to filter after the fact.

I ran the existing benchmark to confirm that this update does not regress performance, even in the case of no filtering being applied. I will post those results in a comment.

**Reviewer Checklist**
- [ ] Implementation matches the proposed design, or proposal is updated to match implementation
- [ ] Sufficient unit test coverage 
- [ ] Sufficient end-to-end test coverage
- [ ] Docs updated or added to `/docs` 
- [ ] Commit messages sensible and descriptive


<!--

Note: If this PR is fixing an issue make sure to add a note saying:
Closes #<ISSUE_NUMBER>

-->