@joelanford 2024-03-26
Here's the promised `benchstat` output comparing `master` and this PR.

```
goos: darwin
goarch: arm64
                      │ master.out  │           filter-fbc.out           │
                      │   sec/op    │   sec/op     vs base               │
LoadFS/1_routines-10    718.5m ± 0%   717.7m ± 0%       ~ (p=0.353 n=10)
LoadFS/10_routines-10   110.7m ± 2%   110.5m ± 1%       ~ (p=0.684 n=10)
LoadFS/20_routines-10   110.9m ± 0%   110.9m ± 1%       ~ (p=0.971 n=10)
geomean                 206.6m        206.4m       -0.09%

                      │  master.out  │           filter-fbc.out            │
                      │     B/op     │     B/op      vs base               │
LoadFS/1_routines-10    234.8Mi ± 0%   233.6Mi ± 0%  -0.49% (p=0.000 n=10)
LoadFS/10_routines-10   237.7Mi ± 0%   236.6Mi ± 0%  -0.43% (p=0.000 n=10)
LoadFS/20_routines-10   237.1Mi ± 0%   236.0Mi ± 0%  -0.47% (p=0.000 n=10)
geomean                 236.5Mi        235.4Mi       -0.47%

                      │ master.out  │           filter-fbc.out           │
                      │  allocs/op  │  allocs/op   vs base               │
LoadFS/1_routines-10    2.561M ± 0%   2.559M ± 0%  -0.09% (p=0.000 n=10)
LoadFS/10_routines-10   2.562M ± 0%   2.559M ± 0%  -0.09% (p=0.000 n=10)
LoadFS/20_routines-10   2.561M ± 0%   2.559M ± 0%  -0.09% (p=0.000 n=10)
geomean                 2.561M        2.559M       -0.09%
```